# api/swagger.yaml
...
swagger: "2.0"
info:
  description: |
    Portainer API is an HTTP API served by Portainer. It is used by the Portainer UI and everything you can do with the UI can be done using the HTTP API.
    Examples are available at https://gist.github.com/deviantony/77026d402366b4b43fa5918d41bc42f8
    You can find out more about Portainer at [http://portainer.io](http://portainer.io) and get some support on [Slack](http://portainer.io/slack/).

    # Authentication

    Most of the API endpoints require to be authenticated as well as some level of authorization to be used.
    Portainer API uses JSON Web Token to manage authentication and thus requires you to provide a token in the **Authorization** header of each request
    with the **Bearer** authentication mechanism.

    Example:
    ```
    Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MSwidXNlcm5hbWUiOiJhZG1pbiIsInJvbGUiOjEsImV4cCI6MTQ5OTM3NjE1NH0.NJ6vE8FY1WG6jsRQzfMqeatJ4vh2TWAeeYfDhP71YEE
    ```

    # Security

    Each API endpoint has an associated access policy, it is documented in the description of each endpoint.

    Different access policies are available:
    * Public access
    * Authenticated access
    * Restricted access
    * Administrator access

    ### Public access

    No authentication is required to access the endpoints with this access policy.

    ### Authenticated access

    Authentication is required to access the endpoints with this access policy.

    ### Restricted access

...

...

...

...

...

...

...

...

...
host: "portainer.domain"
basePath: "/api"
...
schemes:
...
paths:
...
  /endpoints:
    get:
      tags:
      - "endpoints"
      summary: "List endpoints"
      description: |
        List all endpoints based on the current user authorizations. Will
        return all endpoints if using an administrator account otherwise it will
        only return authorized endpoints.
        **Access policy**: restricted
      operationId: "EndpointList"
      produces:
      - "application/json"
      security:
      - jwt: []
      parameters: []
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/EndpointListResponse"
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/GenericError"
    post:
      tags:
      - "endpoints"
      summary: "Create a new endpoint"
      description: |
        Create a new endpoint that will be used to manage a Docker environment.
        **Access policy**: administrator
      operationId: "EndpointCreate"
      consumes:
      - "multipart/form-data"
      produces:
      - "application/json"
      security:
      - jwt: []
      parameters:
      - name: "Name"
        in: "formData"
        type: "string"
        description: "Name that will be used to identify this endpoint (example: my-endpoint)"
        required: true
      - name: "EndpointType"
        in: "formData"
        type: "integer"
        description: "Environment type. Value must be one of: 1 (Docker environment), 2 (Agent environment) or 3 (Azure environment)"
        required: true
      - name: "URL"
        in: "formData"
        type: "string"
        description: "URL or IP address of a Docker host (example: docker.mydomain.tld:2375).\
          \ Defaults to local if not specified (Linux: /var/run/docker.sock, Windows: //./pipe/docker_engine)"
      - name: "PublicURL"
        in: "formData"
        type: "string"
        description: "URL or IP address where exposed containers will be reachable.\
          \ Defaults to URL if not specified (example: docker.mydomain.tld:2375)"
      - name: "GroupID"
        in: "formData"
        type: "string"
        description: "Endpoint group identifier. If not specified will default to 1 (unassigned)."
      - name: "TLS"
        in: "formData"
        type: "string"
        description: "Require TLS to connect against this endpoint (example: true)"
      - name: "TLSSkipVerify"
        in: "formData"
        type: "string"
        description: "Skip server verification when using TLS (example: false)"
      - name: "TLSSkipClientVerify"
        in: "formData"
        type: "string"
        description: "Skip client verification when using TLS (example: false)"
      - name: "TLSCACertFile"
        in: "formData"
        type: "file"
        description: "TLS CA certificate file"
      - name: "TLSCertFile"
        in: "formData"
        type: "file"
        description: "TLS client certificate file"
      - name: "TLSKeyFile"
        in: "formData"
        type: "file"
        description: "TLS client key file"
      - name: "AzureApplicationID"
        in: "formData"
        type: "string"
        description: "Azure application ID. Required if endpoint type is set to 3"
      - name: "AzureTenantID"
        in: "formData"
        type: "string"
        description: "Azure tenant ID. Required if endpoint type is set to 3"
      - name: "AzureAuthenticationKey"
        in: "formData"
        type: "string"
        description: "Azure authentication key. Required if endpoint type is set to 3"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/Endpoint"
        400:
          description: "Invalid request"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Invalid request data format"
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/GenericError"
        503:
          description: "Endpoint management disabled"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Endpoint management is disabled"
  /endpoints/{id}:
    get:
      tags:
      - "endpoints"
      summary: "Inspect an endpoint"
      description: |
        Retrieve details abount an endpoint.
        **Access policy**: restricted
      operationId: "EndpointInspect"
      produces:
      - "application/json"
      security:
      - jwt: []
      parameters:
      - name: "id"
        in: "path"
        description: "Endpoint identifier"
        required: true
        type: "integer"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/Endpoint"
        400:
          description: "Invalid request"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Invalid request"
        404:
          description: "Endpoint not found"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Endpoint not found"
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/GenericError"
    put:
      tags:
      - "endpoints"
      summary: "Update an endpoint"
      description: |
        Update an endpoint.
        **Access policy**: administrator
      operationId: "EndpointUpdate"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      security:
      - jwt: []
      parameters:
      - name: "id"
        in: "path"
        description: "Endpoint identifier"
        required: true
        type: "integer"
      - in: "body"
        name: "body"
        description: "Endpoint details"
        required: true
        schema:
          $ref: "#/definitions/EndpointUpdateRequest"
      responses:
        200:
          description: "Success"
        400:
          description: "Invalid request"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Invalid request data format"
        404:
          description: "Endpoint not found"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Endpoint not found"
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/GenericError"
        503:
          description: "Endpoint management disabled"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Endpoint management is disabled"
    delete:
      tags:
      - "endpoints"
      summary: "Remove an endpoint"
      description: |
        Remove an endpoint.
        **Access policy**: administrator
      operationId: "EndpointDelete"
      security:
      - jwt: []
      parameters:
      - name: "id"
        in: "path"
        description: "Endpoint identifier"
        required: true
        type: "integer"
      responses:
        204:
          description: "Success"
        400:
          description: "Invalid request"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Invalid request"
        404:
          description: "Endpoint not found"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Endpoint not found"
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/GenericError"
        503:
          description: "Endpoint management disabled"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Endpoint management is disabled"
...
  /tags:
    get:
      tags:
      - "tags"
      summary: "List tags"
      description: |
        List tags.
        **Access policy**: administrator
      operationId: "TagList"
      produces:
      - "application/json"
      security:
      - jwt: []
      parameters: []
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/TagListResponse"
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/GenericError"
    post:
      tags:
      - "tags"
      summary: "Create a new tag"
      description: |
        Create a new tag.
        **Access policy**: administrator
      operationId: "TagCreate"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      security:
      - jwt: []
      parameters:
      - in: "body"
        name: "body"
        description: "Tag details"
        required: true
        schema:
          $ref: "#/definitions/TagCreateRequest"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/Tag"
        400:
          description: "Invalid request"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Invalid request data format"
        409:
          description: "Conflict"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "A tag with the specified name already exists"
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/GenericError"
...
  /teams/{id}:
    get:
      tags:
      - "teams"
      summary: "Inspect a team"
      description: |
        Retrieve details about a team. Access is only available for administrator and leaders of that team.
        **Access policy**: restricted
      operationId: "TeamInspect"
      produces:
      - "application/json"
      security:
      - jwt: []
      parameters:
      - name: "id"
        in: "path"
        description: "Team identifier"
        required: true
        type: "integer"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/Team"
        400:
          description: "Invalid request"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Invalid request"
        403:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Access denied to resource"
        404:
          description: "Team not found"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Team not found"
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/GenericError"
    put:
      tags:
      - "teams"
      summary: "Update a team"
      description: |
        Update a team.
        **Access policy**: administrator
      operationId: "TeamUpdate"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      security:
      - jwt: []
      parameters:
      - name: "id"
        in: "path"
        description: "Team identifier"
        required: true
        type: "integer"
      - in: "body"
        name: "body"
        description: "Team details"
        required: true
        schema:
          $ref: "#/definitions/TeamUpdateRequest"
      responses:
        200:
          description: "Success"
        400:
          description: "Invalid request"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Invalid request data format"
        404:
          description: "Team not found"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Team not found"
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/GenericError"
    delete:
      tags:
      - "teams"
      summary: "Remove a team"
      description: |
        Remove a team.
        **Access policy**: administrator
      operationId: "TeamDelete"
      security:
      - jwt: []
      parameters:
      - name: "id"
        in: "path"
        description: "Team identifier"
        required: true
        type: "integer"
      responses:
        204:
          description: "Success"
        400:
          description: "Invalid request"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Invalid request"
        404:
          description: "Team not found"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Team not found"
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/GenericError"
...
  /team_memberships:
    get:
      tags:
      - "team_memberships"
      summary: "List team memberships"
      description: |
        List team memberships. Access is only available to administrators and team leaders.
        **Access policy**: restricted
      operationId: "TeamMembershipList"
      produces:
      - "application/json"
      security:
      - jwt: []
      parameters: []
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/TeamMembershipListResponse"
        403:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Access denied to resource"
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/GenericError"
    post:
      tags:
      - "team_memberships"
      summary: "Create a new team membership"
      description: |
        Create a new team memberships. Access is only available to administrators leaders of the associated team.
        **Access policy**: restricted
      operationId: "TeamMembershipCreate"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      security:
      - jwt: []
      parameters:
      - in: "body"
        name: "body"
        description: "Team membership details"
        required: true
        schema:
          $ref: "#/definitions/TeamMembershipCreateRequest"
      responses:
        200:
          description: "Success"
          schema:
            $ref: "#/definitions/TeamMembership"
        400:
          description: "Invalid request"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Invalid request data format"
        403:
          description: "Unauthorized"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Access denied to resource"
        409:
          description: "Team membership already exists"
          schema:
            $ref: "#/definitions/GenericError"
          examples:
            application/json:
              err: "Team membership already exists for this user and team."
        500:
          description: "Server error"
          schema:
            $ref: "#/definitions/GenericError"
...
securityDefinitions:
  jwt:
    type: "apiKey"
    name: "Authorization"
    in: "header"
definitions:
  Tag:
    type: "object"
    properties:
      Id:
        type: "integer"
        example: 1
        description: "Tag identifier"
      Name:
        type: "string"
        example: "org/acme"
        description: "Tag name"
  Team:
    type: "object"
    properties:
      Id:
        type: "integer"
        example: 1
        description: "Team identifier"
      Name:
        type: "string"
        example: "developers"
        description: "Team name"
  TeamMembership:
    type: "object"
    properties:
      Id:
        type: "integer"
        example: 1
        description: "Membership identifier"
      UserID:
        type: "integer"
        example: 1
        description: "User identifier"
      TeamID:
        type: "integer"
        example: 1
        description: "Team identifier"
      Role:
        type: "integer"
        example: 1
        description: "Team role (1 for team leader and 2 for team member)"
...
  TLSConfiguration:
    type: "object"
    properties:
      TLS:
        type: "boolean"
        example: true
        description: "Use TLS"
      TLSSkipVerify:
        type: "boolean"
        example: false
        description: "Skip the verification of the server TLS certificate"
      TLSCACertPath:
        type: "string"
        example: "/data/tls/ca.pem"
        description: "Path to the TLS CA certificate file"
      TLSCertPath:
        type: "string"
        example: "/data/tls/cert.pem"
        description: "Path to the TLS client certificate file"
      TLSKeyPath:
        type: "string"
        example: "/data/tls/key.pem"
        description: "Path to the TLS client key file"
  AzureCredentials:
    type: "object"
    properties:
      ApplicationID:
        type: "string"
        example: "eag7cdo9-o09l-9i83-9dO9-f0b23oe78db4"
        description: "Azure application ID"
      TenantID:
        type: "string"
        example: "34ddc78d-4fel-2358-8cc1-df84c8o839f5"
        description: "Azure tenant ID"
      AuthenticationKey:
        type: "string"
        example: "cOrXoK/1D35w8YQ8nH1/8ZGwzz45JIYD5jxHKXEQknk="
        description: "Azure authentication key"
...
  UserAccessPolicies:
    type: "object"
    description: "User access policies associated to a registry/endpoint/endpoint group. RoleID is not required for registry access policies and can be set to 0."
    additionalProperties:
      $ref: "#/definitions/AccessPolicy"
    example:
      1: { RoleID: 1 }
      2: { RoleID: 3 }
  TeamAccessPolicies:
    type: "object"
    description: "Team access policies associated to a registry/endpoint/endpoint group. RoleID is not required for registry access policies and can be set to 0."
    additionalProperties:
      $ref: "#/definitions/AccessPolicy"
    example:
      1: { RoleID: 1 }
      2: { RoleID: 3 }
  AccessPolicy:
    type: "object"
    properties:
      RoleID:
        type: "integer"
        example: "1"
        description: "Role identifier. Reference the role that will be associated to this access policy"
...
  Endpoint:
    type: "object"
    properties:
      Id:
        type: "integer"
        example: 1
        description: "Endpoint identifier"
      Name:
        type: "string"
        example: "my-endpoint"
        description: "Endpoint name"
      Type:
        type: "integer"
        example: 1
        description: "Endpoint environment type. 1 for a Docker environment, 2 for an agent on Docker environment or 3 for an Azure environment."
      URL:
        type: "string"
        example: "docker.mydomain.tld:2375"
        description: "URL or IP address of the Docker host associated to this endpoint"
      PublicURL:
        type: "string"
        example: "docker.mydomain.tld:2375"
        description: "URL or IP address where exposed containers will be reachable"
      GroupID:
        type: "integer"
        example: 1
        description: "Endpoint group identifier"
      AuthorizedUsers:
        type: "array"
        description: "List of user identifiers authorized to connect to this endpoint"
        items:
          type: "integer"
          example: 1
          description: "User identifier"
      AuthorizedTeams:
        type: "array"
        description: "List of team identifiers authorized to connect to this endpoint"
        items:
          type: "integer"
          example: 1
          description: "Team identifier"
      TLSConfig:
        $ref: "#/definitions/TLSConfiguration"
      AzureCredentials:
        $ref: "#/definitions/AzureCredentials"
  EndpointSubset:
    type: "object"
    properties:
      Id:
        type: "integer"
        example: 1
        description: "Endpoint identifier"
      Name:
        type: "string"
        example: "my-endpoint"
        description: "Endpoint name"
      Type:
        type: "integer"
        example: 1
        description: "Endpoint environment type. 1 for a Docker environment, 2 for an agent on Docker environment, 3 for an Azure environment."
      URL:
        type: "string"
        example: "docker.mydomain.tld:2375"
        description: "URL or IP address of the Docker host associated to this endpoint"
      PublicURL:
        type: "string"
        example: "docker.mydomain.tld:2375"
        description: "URL or IP address where exposed containers will be reachable"
      GroupID:
        type: "integer"
        example: 1
        description: "Endpoint group identifier"
      AuthorizedUsers:
        type: "array"
        description: "List of user identifiers authorized to connect to this endpoint"
        items:
          type: "integer"
          example: 1
          description: "User identifier"
      AuthorizedTeams:
        type: "array"
        description: "List of team identifiers authorized to connect to this endpoint"
        items:
          type: "integer"
          example: 1
          description: "Team identifier"
      TLSConfig:
        $ref: "#/definitions/TLSConfiguration"
  GenericError:
    type: "object"
    properties:
      err:
        type: "string"
        example: "Something bad happened"
        description: "Error message"
...
  EndpointListResponse:
    type: "array"
    items:
      $ref: "#/definitions/EndpointSubset"
...
  EndpointUpdateRequest:
    type: "object"
    properties:
      Name:
        type: "string"
        example: "my-endpoint"
        description: "Name that will be used to identify this endpoint"
      URL:
        type: "string"
        example: "docker.mydomain.tld:2375"
        description: "URL or IP address of a Docker host"
      PublicURL:
        type: "string"
        example: "docker.mydomain.tld:2375"
        description: "URL or IP address where exposed containers will be reachable.\
          \ Defaults to URL if not specified"
      GroupID:
        type: "integer"
        example: "1"
        description: "Group identifier"
      TLS:
        type: "boolean"
        example: true
        description: "Require TLS to connect against this endpoint"
      TLSSkipVerify:
        type: "boolean"
        example: false
        description: "Skip server verification when using TLS"
      TLSSkipClientVerify:
        type: "boolean"
        example: false
        description: "Skip client verification when using TLS"
      ApplicationID:
        type: "string"
        example: "eag7cdo9-o09l-9i83-9dO9-f0b23oe78db4"
        description: "Azure application ID"
      TenantID:
        type: "string"
        example: "34ddc78d-4fel-2358-8cc1-df84c8o839f5"
        description: "Azure tenant ID"
      AuthenticationKey:
        type: "string"
        example: "cOrXoK/1D35w8YQ8nH1/8ZGwzz45JIYD5jxHKXEQknk="
        description: "Azure authentication key"
      UserAccessPolicies:
        $ref: "#/definitions/UserAccessPolicies"
      TeamAccessPolicies:
        $ref: "#/definitions/TeamAccessPolicies"
...
  TagListResponse:
    type: "array"
    items:
      $ref: "#/definitions/Tag"
  TagCreateRequest:
    type: "object"
    required:
    - "Name"
    properties:
      Name:
        type: "string"
        example: "org/acme"
        description: "Name"
...
  TeamUpdateRequest:
    type: "object"
    required:
    - "Name"
    properties:
      Name:
        type: "string"
        example: "developers"
        description: "Name"
...
  TeamMembershipCreateRequest:
    type: "object"
    required:
    - "UserID"
    - "TeamID"
    - "Role"
    properties:
      UserID:
        type: "integer"
        example: 1
        description: "User identifier"
      TeamID:
        type: "integer"
        example: 1
        description: "Team identifier"
      Role:
        type: "integer"
        example: 1
        description: "Role for the user inside the team (1 for leader and 2 for regular member)"
  TeamMembershipListResponse:
    type: "array"
    items:
      $ref: "#/definitions/TeamMembership"
...